IMD 2.6.5

+++ decode.py
@@ -113,25 +113,26 @@ def decode_sweep(dir_scramblers, dir_WAs, differentiate:int,
 
                 else:
                     this_sc_file=os.path.join(dir_scramblers, 'N_'+str(N),  'N_'+str(N)+'_diff_'+str(diff)+'.npz')
-                    this_scrambler=np.load(this_sc_file)['sc']
-                    scrambler.update({diff:this_scrambler})
-                    dpath=os.path.join(Npath,'diff_'+str(diff))
-                    WApath=os.path.join(dpath,'WAs')
-                    filenames = next(os.walk(WApath), (None, None, []))[2]
-                    for fname in filenames:
-                        fdir=os.path.join(WApath,fname)
-                        WA=np.genfromtxt(fdir, delimiter=",")
-                        method=re.sub('^WA_', '', fname)
-                        method=re.sub('_.*$', '', method)
-                        dict_decode=decode_precomp(well_assigner=WA, differentiate=diff, 
+                    if os.path.isfile(this_sc_file):
+                        this_scrambler=np.load(this_sc_file)['sc']
+                        scrambler.update({diff:this_scrambler})
+                        dpath=os.path.join(Npath,'diff_'+str(diff))
+                        WApath=os.path.join(dpath,'WAs')
+                        filenames = next(os.walk(WApath), (None, None, []))[2]
+                        for fname in filenames:
+                            fdir=os.path.join(WApath,fname)
+                            WA=np.genfromtxt(fdir, delimiter=",")
+                            method=re.sub('^WA_', '', fname)
+                            method=re.sub('_.*$', '', method)
+                            dict_decode=decode_precomp(well_assigner=WA, differentiate=diff, 
 
-                        scrambler=scrambler, readout=np.nan, max_differentiate=-1, sweep=True, **kwargs)
-                        decpath=os.path.join(dpath,'decoders')
-                        if not os.path.exists(decpath):
-                            os.makedirs(decpath)
-                        decname=os.path.join(decpath, 'decoder_'+method+'.json')
+                            scrambler=scrambler, readout=np.nan, max_differentiate=-1, sweep=True, **kwargs)
+                            decpath=os.path.join(dpath,'decoders')
+                            if not os.path.exists(decpath):
+                                os.makedirs(decpath)
+                            decname=os.path.join(decpath, 'decoder_'+method+'.json')
 
-                        json.dump( dict_decode, open(decname, 'w' ) )
+                            json.dump( dict_decode, open(decname, 'w' ) )
 
 
                 DTS=np.round((time.time() - start_time),2)
